Fix text duplication in some scenarios in Minecraft 1.8.9 - 1.12.2
Generally, this wasn't an issue for 99% of usecases. But, unfortunately, due to how Hypixel formats message components sent to the game client, the converter would manage to duplicate the contents of messages and obscure the formatting codes of those messages at the same time.

This was caused by the full message component being appended to our text holder during conversion rather than that specific node's content, meaning that if any child nodes existed and/or the content was predominantly contained within them, they'd be duplicated.