Allow an ending string in text formatting
This adds `start` and `end` strings to `TextFormat`, which allows custom formatting to apply an ending string after text content as well as before it.

For backwards compatibility, there's an internal `startValue` property which will return `toString` if no `start` value is provided, effectively falling back to `start` if an implementor is still going by the old contract.