Auto-reset after each text node when using SCOPED collapse mode
When using DELTA, this will **usually** just resolve as a reset code at the very end of your string (so, no visible change). I've observed some scenarios where this isn't the case, and it's appended elsewhere, but usually this is to the user's benefit.

SCOPED mode allows it to reset after each visually formatted text node, simulating the rendered state of Minecraft's components in 1.16.5+. Mostly used when rendering your text or displaying it via the font renderer.